% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicalPlot.R
\name{calibratePlot}
\alias{calibratePlot}
\title{Plot calibre curve of a lasso cox survival model}
\usage{
calibratePlot(
  cal,
  xlim_min = 0,
  xlim_max = 1,
  ylim_min = 0,
  ylim_max = 1,
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{cal}{rms::calibrate object}

\item{xlim_min}{min value of x axis}

\item{ylim_min}{min value of y axis}

\item{ylim_max}{max value of y axis}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}
}
\value{
calibre line plot
}
\description{
Plot calibre curve of a lasso cox survival model
}
\examples{
library(rms)
set.seed(1)
n <- 200
d.time <- rexp(n)
x1 <- runif(n)
x2 <- factor(sample(c("a", "b", "c"), n, TRUE))
f <- cph(Surv(d.time) ~ pol(x1, 2) * x2, x = TRUE, y = TRUE, surv = TRUE, time.inc = 1.5)
cal <- calibrate(f, u = 1.5, cmethod = "KM", m = 50, B = 20)
calibratePlot(cal)
}
\author{
Yangming si
}
