% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clinical}
\alias{clinical}
\title{riskScore data for plot}
\format{
A data frame
\describe{
\item{id}{TCGA patient Identifier}
\item{futime}{survival time}
\item{fustat}{survival status}
\item{age}{How old is patient}
\item{grade}{cancer grade}
\item{stage}{tumor stage}
\item{T}{tumor T stage}
\item{M}{tumor M stage}
\item{N}{tumor N stage}
\item{riskScore}{Cox model predicted result}
}
}
\source{
TCGA-STAD
}
\usage{
data(clinical)
}
\description{
A dataset containing Cox model predicted results
}
\keyword{datasets}
