% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testBatch.R
\name{CorTestBatch}
\alias{CorTestBatch}
\title{Batch compute Correlation between two Dataframe}
\usage{
CorTestBatch(
  df1,
  df2,
  CorrelationCutoff = 0.5,
  PvalueCutoff = 0.05,
  alternative = "two.sided",
  method = "pearson"
)
}
\arguments{
\item{df1}{A data frame}

\item{df2}{A data frame}

\item{CorrelationCutoff}{Correlation Cutoff value}

\item{PvalueCutoff}{Pvalue Cutoff}

\item{alternative}{one of "two.sided", "greater" or "less"}

\item{method}{One of "pearson", "kendall", or "spearman", can be abbreviated}
}
\value{
A tibble of Correlation result
}
\description{
Batch compute Correlation between two Dataframe
}
\examples{
df1 <- data.frame(gene1 = rnorm(50), gene2 = rnorm(50))
df2 <- data.frame(gene3 = rnorm(50), gene4 = rnorm(50))
CorTestBatch(df1, df2)
}
\author{
Yangming si
}
