% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicalModel.R
\name{cox8000}
\alias{cox8000}
\title{unicox multicox model}
\usage{
cox8000(mydata0)
}
\arguments{
\item{mydata}{a data frame of survival data}
}
\value{
cox result table
}
\description{
unicox multicox model
}
\examples{
data(clinical)
x <- as.data.frame(clinical)[-1][1:3]
cox8000(x)
}
\author{
Yangming si
}
