% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSummary.R
\name{data_summary}
\alias{data_summary}
\title{Calculate mean and standard deviation of data frame}
\usage{
data_summary(data, varname, groupnames)
}
\arguments{
\item{data}{A data frame}

\item{varname}{column name of data to calculate}

\item{groupnames}{column name of data represent different group}
}
\value{
mean and sd of groupname
}
\description{
Calculate mean and standard deviation of data frame
}
\examples{
data("clinical")
data_summary(clinical, "futime", "gender")
}
\author{
Yangming si
}
