% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs.R
\name{DF2matrix}
\alias{DF2matrix}
\title{Convert expression data frame to no duplicated matrix}
\usage{
DF2matrix(df, nodups = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{nodups}{logistic value, TRUE value will remove duplicated row in df}
}
\value{
A matrix
}
\description{
Convert expression data frame to no duplicated matrix
}
\examples{
df <- data.frame(gene = c("gene1", "gene2", "gene1", "gene2"),
                 sample1 = c(1,2,3, 5), sample2 = c(11,12,13, 6),
                 sample1 = c(7, 8, 9, 9), sample2 = c(4, 5, 6, 10))
DF2matrix(df)
}
\author{
Yangming si
}
