% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSummary.R
\name{.summary_func}
\alias{.summary_func}
\title{Calculate mean and standard deviation}
\usage{
.summary_func(x, col)
}
\arguments{
\item{x}{A list contain values}

\item{col}{Name of sub-list}
}
\value{
A vector of mean and standard deviation
}
\description{
Calculate mean and standard deviation
}
\examples{
See data_summary function
}
\author{
Yangming si
}
