% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{filter_CIBERSORT}
\alias{filter_CIBERSORT}
\title{filter CIBERSORT Result}
\usage{
filter_CIBERSORT(results, percent = 0.5)
}
\arguments{
\item{results}{CIBERSORT Result matrix}

\item{percent}{remain A column in a vector containing a proportion of zero less than 'percent', default 0.5}

\item{CorrelationCutoff}{Correlation Cutoff value}

\item{PvalueCutoff}{Pvalue Cutoff}

\item{alternative}{one of "two.sided", "greater" or "less"}

\item{method}{One of "pearson", "kendall", or "spearman", can be abbreviated}
}
\value{
filtered matrix
}
\description{
filter CIBERSORT Result
}
\examples{
library(CIBERSORT)
data(LM22)
data(mixed_expr)
results <- cibersort(sig_matrix = LM22, mixture_file = mixed_expr)
filtered_results <- filter_CIBERSORT(results)
}
\author{
Yangming si
}
