% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAexpression.R
\name{fpkmToTpm}
\alias{fpkmToTpm}
\title{Convert Fpkm matrix to Tpm expression}
\usage{
fpkmToTpm(fpkm)
}
\arguments{
\item{fpkm}{Fpkm count matrix from NGS}
}
\value{
A Tpm expression matrix
}
\description{
Convert Fpkm matrix to Tpm expression
}
\examples{
cnts <- c(4250, 3300, 200, 1750, 50, 0)
lens <- c(900, 1020, 2000, 770, 3000, 1777)
countDf <- data.frame(count = cnts, length = lens)
# assume a mean(FLD) = 203.7
countDf$effLength <- countDf$length - 203.7 + 1
countDf$fpkm <- with(countDf, countToFpkm(count, effLength))
with(countDf, all.equal(tpm, fpkmToTpm(fpkm)))
}
\author{
Yangming si
}
