% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs.R
\name{GetArrayDEGs}
\alias{GetArrayDEGs}
\title{Differential expression analysis on microarrays using limma}
\usage{
GetArrayDEGs(
  mat,
  contrasts,
  design,
  logFCcutoff = 1,
  PvalueCutoff = 0.05,
  group = "Two"
)
}
\arguments{
\item{mat}{Gene expression matrix}

\item{contrasts}{Set comparison group, e.g. treat group - control group}

\item{design}{group design matrix from \code{model.matrix()} function}

\item{logFCcutoff}{cutoff value of logFC}

\item{PvalueCutoff}{cutoff value of p}

\item{group}{"Two" or "Multi"}
}
\value{
A Data Frame of DEGs result
}
\description{
Differential expression analysis on microarrays using limma
}
\examples{
library(CLL)
data("CLLbatch")
mat <- exprs(CLLbatch)
mat <- LOG2Transform(mat)
group <- rep(c("Control", "Treat"), each = ncol(mat)/2)
design <- GetDesign(group, mat)
contrasts <- "Treat-Control"
deg  <- GetArrayDEGs(mat, contrasts, design, logFCcutoff = 1, PvalueCutoff = 0.05)
head(deg)
}
\author{
Yangming si
}
