% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs.R
\name{GetDesign}
\alias{GetDesign}
\title{Convert group vector to design matrix}
\usage{
GetDesign(group, mat)
}
\arguments{
\item{group}{character vector of group}
}
\value{
A group design matrix
}
\description{
Convert group vector to design matrix
}
\examples{
df <- data.frame(gene = c("gene1", "gene2", "gene1", "gene2"),
                 sample1 = c(1,2,3, 5), sample2 = c(11,12,13, 6),
                 sample1 = c(7, 8, 9, 9), sample2 = c(4, 5, 6, 10))
mat <- DF2matrix(df)
group <- rep(c("Control", "Treat"), each = 2)
GetDesign(group, mat)
}
\author{
Yangming si
}
