% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggboxplot2}
\alias{ggboxplot2}
\title{Box plot for One vs More Variables: immunity Results}
\usage{
ggboxplot2(col1, col2, mat)
}
\arguments{
\item{col1}{column names of input data frame, only one}

\item{col2}{column names of input data frame, One or More}

\item{mat}{Expression Matrix or data frame, not tibble}
}
\value{
A ggplot2 object of box plot
}
\description{
Box plot for One vs More Variables: immunity Results
}
\examples{
df1 <- data.frame(gene1 = rnorm(50), gene2 = rnorm(50))
df2 <- data.frame(gene3 = rnorm(50), gene4 = rnorm(50))
df <- cbind(df1, df2)
ggboxplot2("gene1", "gene3", df)
}
\author{
Yangming si
}
