% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggscatter2}
\alias{ggscatter2}
\title{Scatter plot for immunity Results}
\usage{
ggscatter2(genename, cell, gene_mat, cell_matrix, trend = "pos")
}
\arguments{
\item{genename}{gene Name}

\item{cell}{Cell Name}

\item{gene_mat}{Gene Expression Matrix}

\item{cell_matrix}{Cell signal Matrix}

\item{trend}{pos or neg}
}
\value{
A ggplot2 object of scatter plot
}
\description{
Scatter plot for immunity Results
}
\examples{
df1 <- data.frame(gene1 = rnorm(50), gene2 = rnorm(50))
df2 <- data.frame(gene3 = rnorm(50), gene4 = rnorm(50))
ggscatter2("gene1", "gene3", df1, df2)
}
\author{
Yangming si
}
