% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Group2Comparison}
\alias{Group2Comparison}
\title{Covert Group vector to a comparison list}
\usage{
Group2Comparison(group)
}
\arguments{
\item{group}{group character vector}
}
\value{
A group comparison list
}
\description{
Covert Group vector to a comparison list
}
\examples{
group <- as.character(1:3)
Group2Comparison(group)
}
\author{
Yangming si
}
