% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs.R
\name{LOG2Transform}
\alias{LOG2Transform}
\title{Automatic determination of whether to log transform the array expression matrix}
\usage{
LOG2Transform(mat)
}
\arguments{
\item{mat}{Gene expression matrix}
}
\value{
A raw or log-transformed(if needed) Data Frame
}
\description{
Automatic determination of whether to log transform the array expression matrix
}
\examples{
library(CLL)
data("CLLbatch")
mat <- exprs(CLLbatch)
mat1 <- LOG2Transform(mat)
}
\author{
Yangming si
}
