% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSummary.R
\name{mean2}
\alias{mean2}
\title{Calculate average of different group}
\usage{
mean2(x, group = group)
}
\arguments{
\item{x}{A data frame or tibble}

\item{group}{A factor character vector}
}
\value{
A vector of different group average
}
\description{
Calculate average of different group
}
\examples{
x <- c(rnorm(10), runif(10), rnorm(10, 1))
f <- gl(3, 10)
f
mean2(x, f)
}
\author{
Yangming si
}
