% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PCA_new}
\alias{PCA_new}
\title{Plot PCA.}
\usage{
PCA_new(expr, ntop = 500, group, show_name = F)
}
\arguments{
\item{expr}{Gene Expression Matrix}

\item{ntop}{How many genes to use}

\item{group}{group vector}

\item{show_name}{show name or not.}
}
\value{
ggplot object.
}
\description{
PCA Plot for Gene Expression
}
\examples{
library(CLL)
data("CLLbatch")
mat <- exprs(CLLbatch)
group <- rep(c("Control", "Treat"), each = ncol(mat) / 2)
PCA_new(mat, group = group)
}
