% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs.R
\name{pkgs_in}
\alias{pkgs_in}
\title{Batch compute Logistic regression model}
\usage{
pkgs_in(
  pkgs,
  cran_mirror = "https://mirrors.tuna.tsinghua.edu.cn/CRAN/",
  bioc_mirror = "http://mirrors.tuna.tsinghua.edu.cn/bioconductor/"
)
}
\arguments{
\item{pkgs}{Package names to install}

\item{cran_mirror}{CRAN Mirror website}

\item{bioc_mirror}{Bioconductor Mirror website}
}
\description{
Batch compute Logistic regression model
}
\examples{
pkgs <- c("tidyverse", "limma", "affy", "oligo", "lumi")
pkgs_in(pkgs)
}
\author{
Yangming si
}
