% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSummary.R
\name{pValueFormat}
\alias{pValueFormat}
\title{PValue format}
\usage{
pValueFormat(pValue)
}
\arguments{
\item{pValue}{a vector of p value}
}
\value{
formatted p value
}
\description{
PValue format
}
\examples{
library(qvalue)
data(hedenfalk)
p <- hedenfalk$p[1]
pValueFormat(p)
}
\author{
Yangming si
}
