% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riskScore}
\alias{riskScore}
\title{riskScore data for plot}
\format{
A data frame
\describe{
\item{id}{TCGA patient Identifier}
\item{riskScore}{Cox model predicted result}
}
}
\source{
TCGA-STAD
}
\usage{
data(riskScore)
}
\description{
A dataset containing Cox model predicted results
}
\keyword{datasets}
