% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicalPlot.R
\name{RiskScoreRankPlot}
\alias{RiskScoreRankPlot}
\title{Risk Score Rank Plot of a lasso cox survival model}
\usage{
RiskScoreRankPlot(riskScore)
}
\arguments{
\item{riskScore}{a vector of risk Score}
}
\value{
Risk Score Rank plot
}
\description{
Risk Score Rank Plot of a lasso cox survival model
}
\examples{
data(riskScore)
}
\author{
Yangming si
}
