% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicalModel.R
\name{SurvivalModel}
\alias{SurvivalModel}
\title{survival cox model}
\usage{
SurvivalModel(x)
}
\arguments{
\item{x}{a data frame of survival data}
}
\value{
A data frame of survival model
}
\description{
survival cox model
}
\examples{
data(clinical)
x <- as.data.frame(clinical)[-1][1:3]
SurvivalModel(x)
}
\author{
Yangming si
}
