% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testBatch.R
\name{uni_logistic_batch}
\alias{uni_logistic_batch}
\title{Batch compute Logistic regression model}
\usage{
uni_logistic_batch(x, y, dat, PvalueCutoff = 0.05)
}
\arguments{
\item{x}{Independent variable name vector}

\item{y}{Dependent variable name}

\item{dat}{data matrix}
}
\value{
A tibble of Logistic regression result
}
\description{
Batch compute Logistic regression model
}
\examples{
df1 <- data.frame(gene1 = rnorm(50), gene2 = rnorm(50))
df2 <- data.frame(gene3 = rnorm(50), gene4 = rnorm(50))
CorTestBatch(df1, df2)
}
\author{
Yangming si
}
