% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Plot Volcano.}
\usage{
VolcanoPlot(
  DEGs,
  logfc = "logFC",
  pvalue = "P.Value",
  col = NULL,
  xlab = "log2 fold chage",
  ylab = "-log10(p-value)",
  v = c(-1, 1),
  h = -log10(0.05),
  line_col = "green",
  ...
)
}
\arguments{
\item{DEGs}{generated by \code{\link{GetArrayDEGs}}}

\item{logfc}{Log Fold Change}

\item{pvalue}{P value}

\item{col}{vector of colors to choose from.}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{v}{Horizontalism line}

\item{h}{Vertical line}

\item{line_col}{line color}

\item{...}{Other options passed to \code{\link{plot}}}
}
\value{
nothing.
}
\description{
Volcano Plot for DEGs
}
\examples{
library(CLL)
data("CLLbatch")
mat <- exprs(CLLbatch)
mat <- LOG2Transform(mat)
group <- rep(c("Control", "Treat"), each = ncol(mat) / 2)
design <- GetDesign(group, mat)
contrasts <- "Treat-Control"
deg <- GetArrayDEGs(mat, contrasts, design, logFCcutoff = 1, PvalueCutoff = 0.05)
col <- ifelse(deg$P.Value <= 0.05 & abs(deg$logFC) > 1, "blue", "gray")
VolcanoPlot(deg, col = col, main = "Volcano Plot")
colour1 <- ifelse(allDiff$P.Value <= 0.05 & abs(allDiff$logFC) > 1, "green", "gray")
colour2 <- ifelse(allDiff$P.Value > 0.05, "gray", ifelse(allDiff$logFC > 1, "red", ifelse(allDiff$logFC < -1, "green", "gray")))
}
