% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggboxplot2}
\alias{ggboxplot2}
\title{Box plot for immunity Results}
\usage{
ggboxplot2(genename, cell, gene_mat, cell_matrix)
}
\arguments{
\item{genename}{gene Name}

\item{cell}{Cell Name: One or More}

\item{gene_mat}{Gene Expression Matrix}

\item{cell_matrix}{Cell signal Matrix}
}
\value{
A ggplot2 object of box plot
}
\description{
Box plot for immunity Results
}
\examples{
df1 <- data.frame(gene1 = rnorm(50), gene2 = rnorm(50))
df2 <- data.frame(gene3 = rnorm(50), gene4 = rnorm(50))
ggboxplot2("gene1", "gene3", df1, df2)
}
\author{
Yangming si
}
