% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicalPlot.R
\name{SurvivalCurvePlot}
\alias{SurvivalCurvePlot}
\title{Survival curve plot}
\usage{
SurvivalCurvePlot(fit, geneName, pval = pval)
}
\arguments{
\item{fit}{model from survival::survfit}

\item{geneName}{gene Name}

\item{pval}{p Value}
}
\value{
Survival curve plot
}
\description{
Survival curve plot
}
\author{
Yangming si
}
